# 🚀 Performance optimizations for v0.3.0 completion (dev branch)