--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -25,15 +25,18 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+          cache: 'pip'
 
       - name: Install project with dev dependencies
+        timeout-minutes: 10
         run: |
-          python -m pip install --upgrade pip
-          pip install -e .[dev]
+          python -m pip install --upgrade pip --timeout=60
+          pip install -e .[dev] --timeout=300 --retries=3
 
       - name: Run tests with coverage
+        timeout-minutes: 15
         run: |
-          pytest --cov=kinda --cov-report=term-missing tests/
+          pytest --cov=kinda --cov-report=term-missing tests/ --tb=short
 
       - name: Test kinda CLI commands
         run: |