--- a/.gitignore
+++ b/.gitignore
@@ -213,3 +213,6 @@ cython_debug/
 marimo/_static/
 marimo/_lsp/
 __marimo__/
+
+# Profiling output files
+*.prof