--- a/examples/python/individual/welp_example.py.knda
+++ b/examples/python/individual/welp_example.py.knda
@@ -49,7 +49,7 @@ def get_score():
         return None
     return random.randint(70, 100)
 
-final_score = get_score() ~welp 50~ish
+final_score = get_score() ~welp (50 + (5 ~ish))
 ~sorta print("Final score:", final_score)
 
 # Welp with list operations
@@ -62,4 +62,4 @@ preferences = get_user_preferences() ~welp ["default_setting"]
 ~sorta print("User preferences:", preferences)
 
 ~sorta print("=== End Demo ===")
-~sorta print("✅ ~welp provides graceful fallbacks for unreliable operations!")
+~sorta print("[+] ~welp provides graceful fallbacks for unreliable operations!")